ISSUE-181: udpated kotlin to 2.1.20, latest version (#197)
Also bumped mockito lib, since we were using 4.x.x mockito-kotin
and that requires mockito-inline, however, mockito-inline is not
compatable with java version 21 due to a lib they depends on. So
we bumped mockito-kotlin to 5.x.x, where it include mockitio-inline
as a part of the core. Where the imcompatable issue is resolved
by itself due to this bump.