ISSUE-207: fix unable to login on release 0.4.6 (#208)
Fixed a bug where user is not able to login to the app on pro guard enabled build.
Whitelisted all classes under `:model` so that they are not obfuscated. Updated 
` :wiring` module with consumer pro guard rules to exclude `:model` classes. 
(cannot do it in :model since that is a java lib)